+++ configs/datasets/culane.yml
@@ -2,7 +2,7 @@ metric: CULaneMetric
 num_classes: 5 # 4 lanes + background
 
 cut_height: &cut_height 270
-dataset_dir: &dataset_dir /home/yinkh/mycode/data/CULane
+dataset_dir: &dataset_dir dataset/culane
 
 TrainDataset:
   name: CULaneDataSet
